Reorganize project structure with modular layout and runtime artifacts
- Move code to src/vpn_tools/ package with ui/ and data/ subdirectories
- Move documentation to docs/ directory (CHANGELOG.md, MAINTENANCE.md, PROJECT_AUDIT.md)
- Move tests to tests/unit/ with proper conftest.py for imports
- Create runtime/ directory for all logs, database, and cache files
- Add backward-compatible wrapper scripts at root for existing usage
- Update all imports to use relative paths within package
- Update CI workflow and documentation to reflect new structure
- Configure logging after runtime directory creation
- Keep root directory clean with only essential files (LICENSE, README, requirements, wrappers)

This improves project maintainability, makes it easier to package, and keeps
the repository organized with clear separation between code, docs, tests, and
runtime artifacts.
