+++ mullvad_coordinates.py
@@ -1,51 +1,12 @@
 #!/usr/bin/env python3
+"""Backward-compatible import bridge for coordinate lookups."""
 
-"""
-Database of correct coordinates for Mullvad server locations.
-This is used instead of relying on potentially incorrect coordinates from Mullvad's output.
-"""
-
-import json
-import logging
 from pathlib import Path
-from typing import Dict, Tuple
+import sys
 
+ROOT = Path(__file__).resolve().parent
+SRC = ROOT / "src"
+if str(SRC) not in sys.path:
+    sys.path.insert(0, str(SRC))
 
-def _load_coordinates() -> Dict[str, Tuple[float, float]]:
-    """Load coordinates from a JSON file with an in-memory cache."""
-
-    data_file = Path(__file__).with_name("coordinates.json")
-    if not data_file.exists():
-        # Fallback to an example file if provided
-        data_file = Path(__file__).with_name("coordinates.example.json")
-
-    if data_file.exists():
-        try:
-            with data_file.open() as f:
-                data = json.load(f)
-            return {key: tuple(value) for key, value in data.items()}
-        except (json.JSONDecodeError, OSError) as e:
-            logging.error("Failed to load coordinates from %s: %s", data_file, e)
-            pass
-
-    return {}
-
-
-# Cache coordinates at import time
-COORDINATES: Dict[str, Tuple[float, float]] = _load_coordinates()
-
-
-def get_coordinates(city: str, country: str) -> Tuple[float, float]:
-    """
-    Get the correct coordinates for a given city and country.
-
-    Args:
-        city: The city name
-        country: The country name
-
-    Returns:
-        A tuple of (latitude, longitude)
-    """
-    location_key = f"{city}, {country}"
-    return COORDINATES.get(location_key, (0.0, 0.0))
-
+from vpn_tools.mullvad_coordinates import *  # noqa: F401,F403
