+++ src/vpn_tools/__init__.py
@@ -1,0 +1,1 @@
+"""VPN tools package."""
