+++ src/vpn_tools/ui/__init__.py
@@ -1,0 +1,1 @@
+"""UI utilities for vpn_tools."""
